+++ src/main/java/com/kruemblegard/block/VeilgrowthBlock.java
@@ -10,7 +10,7 @@ import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.StateDefinition;
 import net.minecraft.world.level.block.state.properties.BooleanProperty;
 
-public class VeilgrowthBlock extends WayfallSurfaceBlock {
+public class VeilgrowthBlock extends WayfallSurfaceCarpetBlock {
 
     public static final BooleanProperty CHARGED = BooleanProperty.create("charged");